Move listwatch pkg
