+++ security/pkg/k8s/controller/workloadsecret.go
@@ -31,7 +31,7 @@ import (
 
 	"istio.io/istio/pkg/spiffe"
 	k8ssecret "istio.io/istio/security/pkg/k8s/secret"
-	"istio.io/istio/security/pkg/listwatch"
+	"istio.io/istio/pkg/listwatch"
 	caerror "istio.io/istio/security/pkg/pki/error"
 	"istio.io/istio/security/pkg/pki/util"
 	certutil "istio.io/istio/security/pkg/util"
